IoT Gateway: Created "looprw" command
This command is inspired by "loop" and "loopw" commands.

It is intended to simplify capturing SDRAM signal waveforms.
It writes 0xAA..AA to the address, 0x55..55 to the next address,
then reads from the address and reads from the next address.
It does all of this in a loop, adding a small delay between
iterations to allow to distinguish between iterations.